Nguyen Ba Thanh - Add Change Language on Menu, Fix Reponsive Header (Title + Account Setting)
