+++ src/components/Menu/MenuNavigate.jsx
@@ -9,14 +9,18 @@ import Icon, {
     TeamOutlined,
     ProjectOutlined,
     LeftOutlined,
-    RightOutlined
+    RightOutlined,
+    GlobalOutlined
 } from "@ant-design/icons";
 import { Menu } from "antd";
 import logo from "../../assets/Logo.png";
 import { useNavigate } from 'react-router-dom'
-import { useState } from "react";
+import { useState, useEffect, memo } from "react";
 import useViewport from "../../hooks/useViewport";
 import miniLogo from "../../assets/chibi-logo.png";
+import { Dropdown, Space, theme } from 'antd';
+import { useTranslation } from "react-i18next";
+import { DownOutlined } from '@ant-design/icons';
 
 const ZaloSvg = () => (
     <svg width="1em" height="1em" viewBox="0 0 26 29" fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -51,68 +55,7 @@ const PositionIcon = (props) => <Icon component={PositionSvg} {...props} />;
 const TechnologyIcon = (props) => <Icon component={TechnologySvg} {...props} />;
 
 
-const items = [
-    {
-        key: 'dashBoard',
-        label: 'Dashboard',
-        icon: <AppstoreOutlined />,
-    },
-    {
-        key: 'cvManagement',
-        label: 'CV Management',
-        icon: <AuditOutlined />,
-        children: [
-            {
-                key: 'approveCV',
-                label: 'Approve CV',
-            },
-            {
-                key: 'confirmCV',
-                label: 'Confirm CV',
-            },
-        ],
-    },
-    {
-        key: 'listManagement',
-        label: 'List Management',
-        icon: <TeamOutlined />,
-        children: [
-            {
-                key: 'internList',
-                label: 'Intern list',
-            },
-            {
-                key: 'groupList',
-                label: 'Group list',
-            },
-        ],
-    },
-    {
-        key: 'projectManagement',
-        label: 'Project Management',
-        icon: <ProjectOutlined />,
-    },
-    {
-        key: 'positionManagement',
-        label: 'Position Management',
-        icon: <PositionIcon />,
-    },
-    {
-        key: 'technologyManagement',
-        label: 'Technology Management',
-        icon: <TechnologyIcon />,
-    },
-    {
-        key: 'groupZaloManagement',
-        label: 'Group Zalo Management',
-        icon: <ZaloIcon />,
-    },
-    {
-        key: 'settings',
-        label: 'Settings',
-        icon: <SettingOutlined />,
-    }
-];
+
 
 const MenuNavigate = ({ buttonClick }) => {
     // When width of page <= 1024 it will return true for isMobile variable
@@ -122,7 +65,13 @@ const MenuNavigate = ({ buttonClick }) => {
 
     // When clicking on the menu, it will redirect to the corresponding link
     const onClick = (value) => {
-        navigate('/' + value.key)
+        if(value.key == 'vi' || value.key == 'en'){
+            // xử lí chuyển ngôn ngữ khi click ở đây
+            console.log(value.key);
+        }
+        else {
+            navigate('/' + value.key)
+        }
     }
 
     const [collapsed, setCollapsed] = useState(false);
@@ -131,6 +80,137 @@ const MenuNavigate = ({ buttonClick }) => {
     const toggleCollapsed = () => {
         setCollapsed(!collapsed);
         buttonClick()
+    };
+
+    const { useToken } = theme;
+    const { i18n } = useTranslation();
+    const [selectedLanguage, setSelectedLanguage] = useState(i18n.language);
+    const getCurrentLang = () => i18n.language;
+    useEffect(() => {
+        setSelectedLanguage(i18n.language);
+    }, [i18n.language]);
+
+    const handleLanguageChange = (language) => {
+        i18n.changeLanguage(language);
+        setSelectedLanguage(language);
+    };
+
+    const items = [
+        {
+            key: 'dashBoard',
+            label: 'Dashboard',
+            icon: <AppstoreOutlined />,
+        },
+        {
+            key: 'cvManagement',
+            label: 'CV Management',
+            icon: <AuditOutlined />,
+            children: [
+                {
+                    key: 'approveCV',
+                    label: 'Approve CV',
+                },
+                {
+                    key: 'confirmCV',
+                    label: 'Confirm CV',
+                },
+            ],
+        },
+        {
+            key: 'listManagement',
+            label: 'List Management',
+            icon: <TeamOutlined />,
+            children: [
+                {
+                    key: 'internList',
+                    label: 'Intern list',
+                },
+                {
+                    key: 'groupList',
+                    label: 'Group list',
+                },
+            ],
+        },
+        {
+            key: 'projectManagement',
+            label: 'Project Management',
+            icon: <ProjectOutlined />,
+        },
+        {
+            key: 'positionManagement',
+            label: 'Position Management',
+            icon: <PositionIcon />,
+        },
+        {
+            key: 'technologyManagement',
+            label: 'Technology Management',
+            icon: <TechnologyIcon />,
+        },
+        {
+            key: 'groupZaloManagement',
+            label: 'Group Zalo Management',
+            icon: <ZaloIcon />,
+        },
+        {
+            key: 'settings',
+            label: 'Settings',
+            icon: <SettingOutlined />,
+        },
+        {
+            key: 'language',
+            label: 'Language' ,
+            icon: <GlobalOutlined />,
+            children: [
+                {
+                    key: 'vi',
+                    label: 'Việt Nam',
+                },
+                {
+                    key: 'en',
+                    label: 'English',
+                },
+            ],
+        }
+    ];
+
+    const itemsLanguage = [
+        {
+            key: '1',
+            label: (
+                <button onClick={(e) => handleLanguageChange(e.target.value)}>
+                    <option value="en">English</option>
+                </button>
+            ),
+        },
+        {
+            key: '2',
+            label: (
+                <button onClick={(e) => handleLanguageChange(e.target.value)}>
+                    <option value="vi">Tiếng Việt</option>
+                </button>
+            ),
+
+        },
+
+    ];
+    const { token } = useToken();
+    const contentStyle = {
+        backgroundColor: token.colorBgElevated,
+        borderRadius: token.borderRadiusLG,
+        boxShadow: token.boxShadowSecondary,
+    };
+    const menuStyle = {
+        boxShadow: 'none',
+    };
+    const languageData = {
+        en: {
+            flag: "https://cdn.builder.io/api/v1/image/assets/TEMP/ed3550ad056d60b4375fcbb86733c2b3eb54425ccccc2358fcad0c028c501dc1?apiKey=41832340d6f545c2a0509736ad9e1693&",
+            abbreviation: "EN",
+        },
+        vi: {
+            flag: "https://th.bing.com/th/id/R.751d701d99ca9f36a2ec10d886db35bb?rik=wL0QNJdgvG%2fQsQ&riu=http%3a%2f%2fimages5.fanpop.com%2fimage%2fphotos%2f25700000%2fvietnamese-flag-vietnamese-places_mina_kimngan-25790392-1024-768.jpg&ehk=Ul%2fE21vkUUol5RlgNc%2b1OZxq5%2fHqVQmwnw42bVUaACU%3d&risl=&pid=ImgRaw&r=0",
+            abbreviation: "VI",
+        },
     };
 
     return (
@@ -152,6 +232,39 @@ const MenuNavigate = ({ buttonClick }) => {
                 <div onClick={toggleCollapsed} className="button-collapsed">
                     {collapsed ? <RightOutlined /> : <LeftOutlined />}
                 </div>
+                <div className="change-language-navigation">
+                    {/* {!collapsed ?
+                        <Dropdown
+                            menu={{
+                                items: itemsLanguage,
+                            }}
+                            dropdownRender={(menu) => (
+                                <div style={contentStyle}>
+                                    {React.cloneElement(menu, {
+                                        style: menuStyle,
+                                    })}
+                                </div>
+                            )}
+                        >
+                            <a onClick={(e) => e.preventDefault()}>
+                                <Space>
+                                    <img
+                                        src={languageData[selectedLanguage].flag}
+                                        alt="Language Flag"
+                                        className="language-flag"
+                                    />
+                                    <span>{languageData[selectedLanguage].abbreviation}</span>
+                                    <a onChange={(e) => handleLanguageSwtich(e.target.value)}>
+                                        {getCurrentLang() === 'en'}
+                                    </a>
+                                    <DownOutlined />
+                                </Space>
+                            </a>
+                        </Dropdown>
+                        :
+                        <div></div>} */}
+
+                </div>
             </div>) : (<div>
                 <Menu
                     onClick={onClick}
@@ -170,4 +283,4 @@ const MenuNavigate = ({ buttonClick }) => {
     );
 };
 
-export default MenuNavigate;
+export default memo(MenuNavigate);
