+++ src/components/Navigation/Navigation.css
@@ -27,6 +27,13 @@
     position: relative;
 }
 
+.content-navigation .header-navigation .account-setting-mobile {
+    width: 100px;
+    height: 100%;
+    margin-right: 50px;
+    position: relative;
+}
+
 .content-navigation .header-navigation .group-header-navigation {
     display: flex;
     height: 100%;
