+++ src/pages/TechnologyManagement/TechnologyManagement.css
@@ -50,6 +50,13 @@
     position: relative;
 }
 
+.content-navigation-technology-management .header-navigation-technology-management .account-setting-technology-management-mobile {
+    width: 100px;
+    height: 100%;
+    margin-right: 50px;
+    position: relative;
+}
+
 .content-navigation-technology-management .navigation-technology-management {
     width: 100%;
     height: 60px;
